# Conflicts:
#	apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/claim/model/common/ClaimLineNdc.java
#	apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/claim/model/institutional/AdjudicationChargeInstitutionalSharedSystems.java
#	apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/claim/model/institutional/AdjudicationInstitutionalRegular.java
#	apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/claim/model/institutional/AdjudicationPpsCms.java
#	apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/claim/model/institutional/ClaimLineAdjudicationChargeInstitutionalNch.java
#	apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/claim/model/institutional/ClaimLineAdjudicationChargeInstitutionalSharedSystems.java
#	apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/claim/model/professional/AdjudicationProfessionalSharedSystems.java
#	apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/claim/model/professional/ClaimLineAdjudicationChargeProfessionalBase.java
#	apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/claim/model/professional/ClaimLineAdjudicationChargeProfessionalNch.java
#	apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/claim/model/professional/ClaimLineAdjudicationChargeProfessionalSharedSystems.java
#	apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/claim/model/professional/ClaimLineProfessionalCmsNch.java
#	apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/claim/model/professional/entities/ClaimItemProfessionalCmsSharedSystems.java
#	apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/claim/model/professional/entities/ClaimProfessionalCmsNch.java
#	apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/claim/model/rx/ClaimLineAdjudicationRx.java
#	apps/bfd-server-ng/src/test/java/gov/cms/bfd/server/ng/__snapshots__/EobPharmacyIT.snap
#	apps/bfd-server-ng/src/test/java/gov/cms/bfd/server/ng/__snapshots__/EobReadIT.snap
#	apps/bfd-server-ng/src/test/java/gov/cms/bfd/server/ng/__snapshots__/EobSamhsaFilterIT.snap
#	apps/bfd-server-ng/src/test/java/gov/cms/bfd/server/ng/__snapshots__/EobSearchIT.snap

+++ apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/claim/model/common/IdrUnit.java
@@ -1,6 +1,7 @@
 package gov.cms.bfd.server.ng.claim.model.common;
 
 import gov.cms.bfd.server.ng.util.SystemUrls;
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.Optional;
 import lombok.AllArgsConstructor;
@@ -37,7 +38,7 @@ public enum IdrUnit {
     return Arrays.stream(values()).filter(v -> v.idrCode.equals(idrCode)).findFirst();
   }
 
-  SimpleQuantity toFhir(double value) {
+  SimpleQuantity toFhir(BigDecimal value) {
     var quantity = new SimpleQuantity();
     quantity
         .setValue(value)
